Split admin router into admin_jobs + admin_merge sub-routers
admin.py was 322 lines across four loosely related surfaces:
registration tokens, merge candidates, sweep schedulers + auto-backup,
and backup CRUD. The merge handlers and the cron schedulers share no
code with the token CRUD; future agents had to scroll past three
unrelated concerns to find the fourth.

- admin_jobs.py — sweep schedulers (schedule-sweep, schedule-seasonal,
  schedule-nightly) + auto-backup. All four cron-authed endpoints share
  JOBS_CRON_TOKEN; the file also exports enqueue_backup_job (renamed
  from the underscore-prefixed version since admin.py imports it for
  the manual-create endpoint) and the private _enqueue_scheduled_sweep
  helper used only within this file.
- admin_merge.py — merge-candidate review (list / merge / dismiss).
  Router-level admin dep so each handler drops the per-endpoint Depends.
- admin.py — slimmed to registration tokens + backups CRUD, mounts the
  two sub-routers under /admin via include_router.

test_admin_nightly.py's monkeypatch now targets admin_jobs.datetime
since the schedule_nightly handler lives there.
